call into hotreload within configprovider on call back
