+++ src/Core/Configurations/RuntimeConfigProvider.cs
@@ -9,7 +9,6 @@ using Azure.DataApiBuilder.Config;
 using Azure.DataApiBuilder.Config.NamingPolicies;
 using Azure.DataApiBuilder.Config.ObjectModel;
 using Azure.DataApiBuilder.Service.Exceptions;
-using Microsoft.Extensions.Options;
 
 namespace Azure.DataApiBuilder.Core.Configurations;
 
@@ -26,8 +25,6 @@ namespace Azure.DataApiBuilder.Core.Configurations;
 /// </remarks>
 public class RuntimeConfigProvider
 {
-    private readonly IOptionsMonitor<RuntimeOptions> _runtimeConfigMonitor; // may not be needed unless runtime updates after initital reload are required
-
     public delegate Task<bool> RuntimeConfigLoadedHandler(RuntimeConfigProvider sender, RuntimeConfig config);
 
     public List<RuntimeConfigLoadedHandler> RuntimeConfigLoadedHandlers { get; } = new List<RuntimeConfigLoadedHandler>();
@@ -47,10 +44,9 @@ public class RuntimeConfigProvider
 
     private RuntimeConfig? _runtimeConfig;
 
-    public RuntimeConfigProvider(RuntimeConfigLoader runtimeConfigLoader, IOptionsMonitor<RuntimeOptions> runtimeConfigMonitor)
+    public RuntimeConfigProvider(RuntimeConfigLoader runtimeConfigLoader)
     {
         ConfigLoader = runtimeConfigLoader;
-        _runtimeConfigMonitor = runtimeConfigMonitor;
     }
 
     /// <summary>
@@ -111,6 +107,11 @@ public class RuntimeConfigProvider
     {
         runtimeConfig = _runtimeConfig;
         return _runtimeConfig is not null;
+    }
+
+    public void HotReloadConfig()
+    {
+        ConfigLoader.TryLoadKnownConfig(out _runtimeConfig);
     }
 
     /// <summary>
